Separated extraction and detection subpackages. Consolidated all of detection's functions to just one call in extraction.
