+++ suite2p/detection/__init__.py
@@ -1,0 +1,1 @@
+from .detect import roi_stats
