+++ suite2p/detection/utils.py
@@ -1,0 +1,51 @@
+import numpy as np
+import math
+
+def fitMVGaus(y,x,lam,thres=2.5):
+    """ computes 2D gaussian fit to data and returns ellipse of radius thres standard deviations.
+
+    Parameters
+    ----------
+    y : float, array
+        pixel locations in y
+    x : float, array
+        pixel locations in x
+    lam : float, array
+        weights of each pixel
+
+    Returns
+    -------
+        mu : float
+            mean of gaussian fit.
+        cov : float
+            covariance of gaussian fit.
+        radii : float, array
+            half of major and minor axis lengths of elliptical fit.
+        ellipse : float, array
+            coordinates of elliptical fit.
+        area : float
+            area of ellipse.
+
+    """
+
+    # normalize pixel weights
+    lam /= lam.sum()
+    # mean of gaussian
+    yx = np.stack((y,x))
+    mu  = (lam*yx).sum(axis=-1)
+    yx = yx - np.expand_dims(mu, axis=1)
+    yx = yx * lam**.5
+    #yx  = np.concatenate((y*lam**0.5, x*lam**0.5),axis=0)
+    cov = yx @ yx.transpose()
+    # radii of major and minor axes
+    radii,evec  = np.linalg.eig(cov)
+    radii = np.maximum(0, np.real(radii))
+    radii       = thres * radii**.5
+    # compute pts of ellipse
+    npts = 100
+    p = np.expand_dims(np.linspace(0, 2*math.pi, npts),axis=1)
+    p = np.concatenate((np.cos(p), np.sin(p)),axis=1)
+    ellipse = (p * radii) @ evec.transpose() + mu
+    area = (radii[0] * radii[1])**0.5 * math.pi
+    radii  = np.sort(radii)[::-1]
+    return mu, cov, radii, ellipse, area
