+++ suite2p/extraction/__init__.py
@@ -1,3 +1,2 @@
 from .dcnv import preprocess, oasis
-from .extract import extract_traces, detect_and_extract, enhanced_mean_image, roi_stats
-from .masks import create_cell_masks, create_neuropil_masks, get_overlaps
+from .extract import extract_traces, detect_and_extract, enhanced_mean_image
