+++ suite2p/io/nwb.py
@@ -4,6 +4,7 @@ import os
 import numpy as np
 import scipy
 
+from .. import detection
 from .. import extraction
 from .. import run_s2p
 
@@ -42,7 +43,7 @@ def read_nwb(fpath):
             if multiplane:
                 stat[-1]['iplane'] = int(rois[n][0,-2])
         ops = run_s2p.default_ops()
-        stat = extraction.roi_stats(ops, stat)
+        stat = detection.roi_stats(ops, stat)
         if multiplane:
             nplanes = np.max(np.array([stat[n]['iplane'] for n in range(len(stat))]))+1
         else:
